--- Assets/Scenes/develop.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ecab97a3cd7a81d4aacfa8933fafc8fc
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
