+++ Assets/Scripts/Atamakko/AtamakkoCore.cs
@@ -1,8 +1,9 @@
 using System;
 using System.Collections.Generic;
+using Player;
 using UnityEngine;
 
-namespace Player
+namespace Atamakko
 {
     public class AtamakkoCore : MonoBehaviour, IDamagable
     {
