+++ Assets/Scripts/Atamakko/EasyCPU.cs
