+++ Assets/Scripts/Atamakko/EasyCPU.cs.meta
