+++ Assets/Scripts/Atamakko/EnemyCore.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Atamakko;
 using Card;
 
 namespace Player
