+++ Assets/Scripts/Atamakko/EnemyCore.cs.meta
