+++ Assets/Scripts/Atamakko/IDamagable.cs
