+++ Assets/Scripts/Atamakko/IDamagable.cs.meta
