+++ Assets/Scripts/Atamakko/PlayerCore.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Atamakko;
 using Card;
 using UnityEngine;
