+++ Assets/Scripts/Atamakko/PlayerCore.cs.meta
