--- Assets/Scripts/Card/CardManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3d0ef4c986bee254ab2a8ff129b9e61e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
