--- Assets/Scripts/Card/CardMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b1c8eae250860794db97dccbcfcc7015
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
