+++ Assets/Scripts/Card/CheckSlot.cs
