+++ Assets/Scripts/Card/CheckSlot.cs.meta
