--- Assets/Scripts/Card/DropPlace.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 539b41ad320ce294db7de924a8b5d6a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
