--- Assets/Scripts/Field.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e5e84cbccad154041b2ea7d4465d81dc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
