--- Assets/Scripts/Player/IMobile.cs
@@ -1,7 +1,0 @@
-namespace Player
-{
-    public interface IMobile
-    {
-        void Move(int slotNum);
-    }
-}
