--- Assets/Scripts/Player/IMobile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7fd46ff3d049eaf44866909f423d7f9f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
