--- Assets/Scripts/Player/PlayerAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a960d6eefe45a34c853840a1b70ead3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
