--- Assets/Scripts/Player/PlayerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d120cbb97fc85844e83429cc7451002b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
