--- Assets/Scripts/TestScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0af7bae8c0e579941a69f97fa286d604
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
